SSHS: convert helpers to C++.
